h2: LCCSCF_H2_MANUAL_RXFLOW and refactor txcr
This gives a client connection a way to start client connections
with the
of the default, and precisely rx flow control incoming data by
meting out further tx credit.

Add another minimal example http-client-h2-rxflow demonstrating how
to force a connection's peer's initial budget to transmit to us
and control it during the connection lifetime to restrict the amount
of incoming data we have to buffer.