docs: add bug report template md
The ease of dropping bug reports is making it a bit asymmetrical
in terms of the complexity of even trying to figure out what is
being told, let alone writing a reproducer and trying it.

Ask for more effort from reporters and provide "first aid"
suggestions and filtering in the template, so at least some of
the support requests will get fixed by people doing that
before I need to hear about it.