remove: generic sessions
As far as I know there are no users of this, although it worked
it's basically unmaintainable due to handling the sql and JSON
manually.

Gradually better capabilities have appeared in lws, like
lws_struct abstracting out the sql and JSON, and now generic
JWT... these have been used in Sai to great effect and displaced
the only organic would-be user of this.

There is a better path to do this stuff now and no point keeping
this around.