lws_ptr_diff_size_t
lws_ptr_diff() is a very handy helper reducing the pointer delta to an int
count of bytes... however many times we actually want a size_t count of bytes
since that's the type of the argument we're trying to feed.

Let's establish a variant that naturally issues size_t.